Fix JSON serialization precision in VAL-016 confidence gate recording

+++ src/integration/val016_source_repair_v2.cpp
@@ -484,7 +484,7 @@ public:
     void recordConfidenceGate(float confidence, const std::string& action) {
         nlohmann::json step;
         step["phase"] = "confidence_gate";
-        step["confidence"] = confidence;
+        step["confidence"] = static_cast<double>(confidence);
         step["action"] = action;
         step["timestamp"] = getTimestamp();
         trace_["steps"].push_back(step);
